requirements.txt
websocket-client
pandas
ta
requests
